[~] Auth core moved to Reception.Core.Auth
