+++ Reception.Core.Auth/Model/AuthenticateRequest.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace Reception.Server.Auth.Models
+namespace Reception.Core.Auth.Model
 {
     public class AuthenticateRequest
     {
