+++ Reception.sln
@@ -38,7 +38,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Reception.Server.Auth.Passw
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{AD2A5F91-93CF-42D8-A331-6CA299E9097E}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Reception.App.Service", "src\Reception.App.Service\Reception.App.Service.csproj", "{91529021-D067-4385-9CF1-DF25A62D7560}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Reception.App.Service", "src\Reception.App.Service\Reception.App.Service.csproj", "{91529021-D067-4385-9CF1-DF25A62D7560}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Reception.Core.Auth", "Reception.Core.Auth\Reception.Core.Auth.csproj", "{CD6D113E-398E-4002-956B-ECFB897470EB}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -98,6 +100,10 @@ Global
 		{91529021-D067-4385-9CF1-DF25A62D7560}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{91529021-D067-4385-9CF1-DF25A62D7560}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{91529021-D067-4385-9CF1-DF25A62D7560}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CD6D113E-398E-4002-956B-ECFB897470EB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CD6D113E-398E-4002-956B-ECFB897470EB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CD6D113E-398E-4002-956B-ECFB897470EB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CD6D113E-398E-4002-956B-ECFB897470EB}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -116,6 +122,7 @@ Global
 		{E6108DCA-A26C-4D35-A41F-2589FAA823BA} = {AD2A5F91-93CF-42D8-A331-6CA299E9097E}
 		{5DDD5B34-5808-4EB1-A491-5DFCC7C239A5} = {AD2A5F91-93CF-42D8-A331-6CA299E9097E}
 		{91529021-D067-4385-9CF1-DF25A62D7560} = {AD2A5F91-93CF-42D8-A331-6CA299E9097E}
+		{CD6D113E-398E-4002-956B-ECFB897470EB} = {AD2A5F91-93CF-42D8-A331-6CA299E9097E}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {D9C0D531-28FB-4D40-AFD0-658855890982}
