+++ src/Reception.Server.Auth/Reception.Server.Auth.csproj
@@ -18,6 +18,7 @@
 	</ItemGroup>
 
 	<ItemGroup>
+		<ProjectReference Include="..\..\Reception.Core.Auth\Reception.Core.Auth.csproj" />
 		<ProjectReference Include="..\Reception.Extensions\Reception.Extension.csproj" />
 		<ProjectReference Include="..\Reception.Model.Interfaces\Reception.Model.Interface.csproj" />
 		<ProjectReference Include="..\Reception.Model.Network\Reception.Model.Network.csproj" />
