+++ src/Reception.Server.Auth/Startup.cs
@@ -5,7 +5,9 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Microsoft.OpenApi.Models;
-using Reception.Server.Auth.Helpers;
+using Reception.Core.Auth.Helpers;
+using Reception.Core.Auth.Logic;
+using Reception.Core.Auth.Repository;
 using Reception.Server.Auth.Logic;
 using Reception.Server.Auth.PasswordHelper;
 using Reception.Server.Auth.Repository;
