+++ src/Reception.Server.Data/Controllers/PersonController.cs
@@ -1,5 +1,4 @@
-﻿using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Mvc;
 using Reception.Model.Dto;
 using Reception.Model.Interface;
 using Reception.Model.Network;
@@ -10,7 +9,6 @@ using System.Threading.Tasks;
 
 namespace Reception.Server.Data.Controllers
 {
-    [Authorize]
     [Route("api/[controller]")]
     [ApiController]
     public class PersonController : ControllerBase, IBaseController
